refactor: useGetDndContextSensor로 로직 분리
