+++ frontend/techpick/src/hooks/index.ts
@@ -1,2 +1,3 @@
 export { useFolderToFolderDndMonitor } from './useFolderToFolderDndMonitor';
 export { usePickToPickDndMonitor } from './usePickToPickDndMonitor';
+export { useGetDndContextSensor } from './useGetDndContextSensor';
