+++ frontend/techpick/src/hooks/useGetDndContextSensor.ts
@@ -1,0 +1,18 @@
+import { MouseSensor, TouchSensor, useSensor, useSensors } from '@dnd-kit/core';
+
+export function useGetDndContextSensor() {
+  const mouseSensor = useSensor(MouseSensor, {
+    activationConstraint: {
+      distance: 10, // MouseSensor: 10px 이동해야 드래그 시작
+    },
+  });
+  const touchSensor = useSensor(TouchSensor, {
+    activationConstraint: {
+      delay: 250,
+      tolerance: 5,
+    },
+  });
+  const sensors = useSensors(mouseSensor, touchSensor);
+
+  return { sensors };
+}
